requirements.txt
PythonTwitchBotFramework
pygame
